musicdb/requirements.txt
apturl==0.5.2
beautifulsoup4==4.4.1
blinker==1.3
Brlapi==0.6.4
certifi==2018.1.18
chardet==3.0.4
checkbox-support==0.22
command-not-found==0.3
cryptography==1.2.3
defer==1.0.6
dj-database-url==0.4.2
Django==2.0.2
django-filter==1.1.0
djangorestframework==3.7.7
feedparser==5.1.3
flake8==3.5.0
guacamole==0.9.2
gunicorn==19.7.1
html5lib==0.999
httplib2==0.9.1
idna==2.6
Jinja2==2.8
language-selector==0.1
louis==2.6.4
lxml==3.5.0
Mako==1.0.3
Markdown==2.6.11
MarkupSafe==0.23
mccabe==0.6.1
oauthlib==1.0.3
onboard==1.2.0
padme==1.1.1
pew==1.1.2
pexpect==4.0.1
Pillow==3.1.2
pipenv==9.0.3
plainbox==0.25
psycopg2-binary==2.7.4
ptyprocess==0.5
pyasn1==0.1.9
pycodestyle==2.3.1
pycups==1.9.73
pycurl==7.43.0
pyflakes==1.6.0
pygobject==3.20.0
PyJWT==1.3.0
pyparsing==2.0.3
python-apt==1.1.0b1
python-debian==0.1.27
python-systemd==231
pytz==2018.3
pyxdg==0.25
reportlab==3.3.0
requests==2.18.4
sessioninstaller==0.0.0
six==1.10.0
sqlparse==0.1.18
system-service==0.3
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.22
usb-creator==0.3.0
virtualenv==15.1.0
virtualenv-clone==0.2.6
whitenoise==3.3.1
xdiagnose==3.8.4.1
xkit==0.0.0
XlsxWriter==0.7.3
